Refactor listing media component and add remove image feature
